requirements.txt
pytest>=8.0.0
pygame-ce
